requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.28
asyncpg==0.29.0
pydantic==2.6.4
pydantic-settings==2.2.1
python-dotenv==1.0.1
cryptography==42.0.5
httpx==0.27.0
# For HF Integration
huggingface-hub==0.21.4
# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
